refactor: 입력된 자동차 이름, 시도할 횟수를 검증하는 함수를 util모듈으로 분리
